+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Unreleased
 
 ##### Added
-* **\[core\]** The `ObservableChangedEvent` spy event will now emit the value of the observable if 
+* **\[core\]** The `ObservableChangedEvent` spy event will now emit the value of the observable if
   property introspectors are enabled and the observable has an accessor introspector.
 * **\[browser-extras\]** Enhance the `ConsoleSpyEventProcessor` to log the value field of the
   `ObservableChangedEvent` spy event if it has been supplied.