+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.29](https://github.com/arez/arez/tree/v0.29) (2017-11-28)
+[Full Changelog](https://github.com/arez/arez/compare/v0.28...v0.29)
 
 ##### Added
 * **\[core\]** The `ObservableChangedEvent` spy event will now emit the value of the observable if